+++ wp-dispensary.php
@@ -11,7 +11,7 @@
  * Plugin Name:       WP Dispensary
  * Plugin URI:        http://www.wpdispensary.com
  * Description:       The complete medical marijuana dispensary menu solution for WordPress.
- * Version:           1.3.1
+ * Version:           1.4
  * Author:            WP Dispensary
  * Author URI:        http://www.wpdispensary.com
  * License:           GPL-2.0+