+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # Changelog
 
 ### 1.4
-* PHP Code Sniffer WordPress standards updates throughout plugin files
+* PHP CodeSniffer WordPress Standards updates throughout various plugin files
+* Created the Topicals menu type, with widget, shortcode and WP-API output options
+* Updated the output style of the shortcode item information
+* Added the WP Dispensary icon to each CPT in the admin dashboard for easier recognition when you're in the admin dashboard
+* Added `active_plugins` check for "Subtitles" plugin in `admin/wp-dispensary-rest-api.php` API output
+* Added sanitization to $_POST in `admin/post-types/wp-dispensary-metaboxes.php` pre-roll flower selection drop down
 
 ### 1.3.1
 * Fixed CBD being called BCD #typingtoofast